refactor: 전공역량 별 교과목 데이터 파싱 함수 분리
